change file structure and done sectionB
